requirements.txt
tensorflow>=2.9
numpy
pandas
scikit-learn
matplotlib
Pillow
flask

kaggle

